Fix weird toolbar styles in demo
This was happening because the buildLoader job was not being passed a
direction, so it was putting in both LTR and RTL styles and relying on
something to disable the RTL styles, but that wasn't happening.

Instead, pass dir: 'ltr' into buildLoader so that it only outputs the
LTR styles.